--- a/src/main/resources/application.properties
+++ b/src/main/resources/application.properties
@@ -8,6 +8,15 @@ spring.sql.init.data-locations=classpath*:db/${database}/data.sql
 
 # Web
 spring.thymeleaf.mode=HTML
+# Template caching
+spring.thymeleaf.cache=true
+spring.thymeleaf.cache-ttl=3600
+spring.thymeleaf.template-resolver-order=1
+
+# View resolver optimization
+spring.mvc.view.prefix=/templates/
+spring.mvc.view.suffix=.html
+spring.mvc.static-path-pattern=/static/**
 
 # JPA
 spring.jpa.hibernate.ddl-auto=none
@@ -18,6 +27,7 @@ spring.data.flaskdb.uri=http://localhost:27017/feedbacks
 
 # Internationalization
 spring.messages.basename=messages/messages
+spring.messages.cache-duration=3600
 
 # Actuator
 management.endpoints.web.exposure.include=*
@@ -29,3 +39,8 @@ logging.level.org.springframework=INFO
 
 # Maximum time static resources should be cached
 spring.web.resources.cache.cachecontrol.max-age=12h
+spring.web.resources.chain.cache=true
+spring.web.resources.chain.strategy.content.enabled=true
+spring.web.resources.chain.strategy.fixed.enabled=true
+spring.web.resources.chain.strategy.fixed.paths=/**
+spring.web.resources.chain.strategy.fixed.version=v1