--- a/src/main/resources/templates/welcome.html
+++ b/src/main/resources/templates/welcome.html
@@ -3,13 +3,27 @@
 <html xmlns:th="https://www.thymeleaf.org" th:replace="~{fragments/layout :: layout (~{::body},'home')}">
 
   <body>
+    <!-- Cache the welcome message fragment -->
+    <div th:fragment="welcome-message" th:cache="true">
+      <h2 th:text="#{welcome}">Welcome</h2>
+    </div>
 
-    <h2 th:text="#{welcome}">Welcome</h2>
-    <div class="row">
+    <!-- Cache the pets image fragment -->
+    <div th:fragment="pets-image" th:cache="true" class="row">
         <div class="col-md-12">
-          <img class="img-responsive" src="../static/resources/images/pets.png" th:src="@{/resources/images/pets.png}"/>
+          <img class="img-responsive" 
+               src="../static/resources/images/pets.png" 
+               th:src="@{/resources/images/pets.png}"
+               loading="lazy"
+               width="1000"
+               height="400"
+               alt="Pets"/>
         </div>
     </div>
+
+    <!-- Include cached fragments -->
+    <div th:replace="~{::welcome-message}"></div>
+    <div th:replace="~{::pets-image}"></div>
 
   </body>