This PR implements performance optimizations for the PetClinic welcome page to reduce rendering time and improve response times.

Current Performance:
- Median response time: 396.21ms
- P95: 413.19ms
- Rendering takes 67.56% of total response time (~270ms)

Optimizations Implemented:

1. Template Caching:
   - Enabled Thymeleaf template caching
   - Added cache TTL configuration
   - Implemented fragment caching for static content

2. View Resolution:
   - Optimized view resolver chain
   - Added template location caching
   - Configured static resource handling

3. Controller Optimizations:
   - Added response caching
   - Implemented HTTP caching headers
   - Optimized view resolution

4. Template Improvements:
   - Added fragment caching
   - Optimized image loading with lazy loading
   - Added image dimensions to prevent layout shifts

Expected Improvements:
- Reduced rendering time to < 100ms
- Total response time target: < 200ms
- Improved caching for static content
- Better resource utilization

Testing:
Please verify the performance improvements by monitoring the response times and rendering metrics after deployment.